Add canary for potential gap in info generation
I am worried that there is a case where OK_to_upload is True
but no description is generated when:
"Beskrivning" is None
"Ort, foto" is not None
any of "Motivord", "Händelse / var närvarande vid" is None

If this scenario ever occurs then a message will be outputted to
the user. To make it more obviouse I've commented out some now
unneded print statements.